Controllo per creazione utenti con lo stesso username

+++ modules/utenti/actions.php
@@ -11,7 +11,7 @@ switch (filter('op')) {
     case 'add':
         $nome = filter('nome');
 
-        // Verifico che questo username non sia già stato usato
+        // Verifico che questo nome gruppo non sia già stato usato
         if ($dbo->fetchNum('SELECT nome FROM zz_groups WHERE nome='.prepare($nome)) == 0) {
             $dbo->query('INSERT INTO zz_groups( nome, editable ) VALUES('.prepare($nome).', 1)');
             flash()->info(tr('Gruppo aggiunto!'));
@@ -42,39 +42,44 @@ switch (filter('op')) {
         $password = filter('password');
 
         $id_utente = filter('id_utente');
-        if (!empty($id_utente)) {
-            $utente = User::find($id_utente);
+        if ($dbo->fetchNum('SELECT username FROM zz_users WHERE id != '.prepare($id_utente).' AND username='.prepare($username)) == 0) {
+            //Aggiunta/modifica utente
+            if (!empty($id_utente)) {
+                $utente = User::find($id_utente);
 
-            $utente->username = $username;
-            $utente->email = $email;
+                $utente->username = $username;
+                $utente->email = $email;
 
-            $cambia_password = filter('change_password');
-            if (!empty($cambia_password)) {
-                $utente->password = $password;
+                $cambia_password = filter('change_password');
+                if (!empty($cambia_password)) {
+                    $utente->password = $password;
+                }
+            } else {
+                $gruppo = \Models\Group::find($id_record);
+                $utente = User::build($gruppo, $username, $email, $password);
             }
-        } else {
-            $gruppo = \Models\Group::find($id_record);
-            $utente = User::build($gruppo, $username, $email, $password);
-        }
 
-        // Foto
-        if (!empty($_FILES['photo']['tmp_name'])) {
-            $utente->photo = $_FILES['photo'];
-        }
+            // Foto
+            if (!empty($_FILES['photo']['tmp_name'])) {
+                $utente->photo = $_FILES['photo'];
+            }
 
-        // Anagrafica
-        $id_anagrafica = filter('idanag');
-        $utente->id_anagrafica = $id_anagrafica;
+            // Anagrafica
+            $id_anagrafica = filter('idanag');
+            $utente->id_anagrafica = $id_anagrafica;
 
-        $utente->save();
+            $utente->save();
 
-        $dbo->query('DELETE FROM zz_user_sedi WHERE id_user = '.prepare($id_utente));
-        $sedi = post('idsede');
-        if (empty($sedi)) {
-            $sedi = [0];
-        }
-        foreach ($sedi as $id_sede) {
-            $dbo->query('INSERT INTO `zz_user_sedi` (`id_user`,`idsede`) VALUES ('.prepare($id_utente).', '.prepare($id_sede).')');
+            $dbo->query('DELETE FROM zz_user_sedi WHERE id_user = '.prepare($id_utente));
+            $sedi = post('idsede');
+            if (empty($sedi)) {
+                $sedi = [0];
+            }
+            foreach ($sedi as $id_sede) {
+                $dbo->query('INSERT INTO `zz_user_sedi` (`id_user`,`idsede`) VALUES ('.prepare($id_utente).', '.prepare($id_sede).')');
+            }
+        }else{
+            flash()->error(tr('Utente già esistente!'));
         }
 
         break;
